requirements.txt
 pandas

 plotly_express
 
 streamlit